requirements.txt
pytest==7.4.3
selenium==4.15.2
allure-pytest==2.13.2
pytest-html==4.1.1
pytest-xdist==3.3.1
webdriver-manager==4.0.1
python-dotenv==1.0.0